requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
APScheduler==3.10.4
qrcode==7.4.2
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
WTForms==3.1.1
